==> 1.3-Data-wrangling.ipynb <==
import pandas as pd

import matplotlib.pyplot as plt
%matplotlib inline

### Load the dataset

# load the dataset
df_transactions = pd.read_csv('data/sample_transactions_with_missing_value.csv', header=0)

# count number of missing values
print('Count number of missing values')
print(df_transactions.isna().sum())

# get a summary of the dataframe with not-null-count for each column
df_transactions.info()

### Handling missing values

# show records with missing value
print('Records with missing unit-price')
df_missing = df_transactions[df_transactions['UnitPrice'].isna() == True]
df_missing.head()


# impute missing values with median of that column
print('Impute missing values with median of that column')
df_imputed = df_missing.fillna(df_transactions['UnitPrice'].median())
df_imputed.head()

### Transposing

# normal descriptive stats
df_transactions.describe()

# transformed descriptive stats
df_transactions.describe().transpose()

### Grouping

# calculate total sale
df_imputed['TotalSale'] = df_imputed['UnitPrice'] * df_imputed['Quantity']

# aggregate total sale per customer
df_summary = df_imputed.groupby('CustomerID').agg({'TotalSale': 'sum', 'InvoiceDate':'first'})
print('Total sale per customer')
df_summary.head()
