==> 1.4-Data-Modeling.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

df_transactions = pd.read_csv('data/mock_transactions.csv', header=0)
df_customers = pd.read_csv('data/mock_customers.csv', header=0)

#### Accessing the datasets

df_transactions.shape

df_transactions.head()  # .tail()

df_customers.shape

df_customers.head()

### Joining

df_combined = df_transactions.merge(right=df_customers, how='left', on='CustomerID')
print(df_combined.shape)
df_combined.head()
