==> 2.1-Advertising.ipynb <==
import pandas as pd
import plotly.express as px

df_advertising = pd.read_csv('data/advertising.csv', index_col=0)
df_advertising.describe()

df_advertising.isna().sum()

fig = px.scatter(df_advertising, x='TV', y='sales', title='TV Advertising Spend vs. Sales')
fig.update_layout(width=800, height=600)
fig.show()

from sklearn.linear_model import LinearRegression

lr = LinearRegression()
lr.fit(df_advertising[['TV']], df_advertising['sales'])

predictions = lr.predict(df_advertising[['TV']])

fig = px.scatter(
    df_advertising,
    x='TV',
    y='sales',
    title='TV Ad. Spend vs. Sales'
)
fig.add_scatter(
    x=df_advertising['TV'],
    y=predictions,
    mode='lines',
    name='Regression Line',
    line=dict(color='red')
)

fig.update_layout(width=800, height=600, showlegend=False)
fig.show()
