==> 2.3-Feature-Engineering.ipynb <==
# import important libraries
import pandas as pd
import datetime as dt
import plotly.express as px
import matplotlib.pyplot as plt
import seaborn as sns

# load the transaction dataset
df_retail = pd.read_csv('data/retail_transactions.csv')

### Data Transformation

# remove unnecessary columns
df_retail = df_retail.drop(columns=['StockCode', 'Description'])

# keep UK records only
df_retail = df_retail[df_retail['Country'] == 'United Kingdom']

# fix the data type and parse datetime
df_retail['CustomerID'] = df_retail['CustomerID'].astype(str)
df_retail['InvoiceDate'] = pd.to_datetime(df_retail['InvoiceDate']).dt.normalize()

# calculate revenue and transaction year
df_retail['Revenue'] = df_retail['UnitPrice'] * df_retail['Quantity']
df_retail['Year'] = df_retail['InvoiceDate'].dt.year

# take a look at the current dataset
df_retail.info()

# take a look at the dataframe
df_retail.head()

# understand the numeric columns
print(df_retail.shape)
df_retail.describe()

### Feature Creation

# First we would like to summarize the total amount of revenue for each invoice.

df_retail = df_retail.groupby('InvoiceNo').agg(
    {
        'Revenue':'sum',
        'CustomerID': 'first',
        'InvoiceDate': 'first',
        'Year': 'first'
    }
)
# print(df_retail.shape)
df_retail.head()
# df_retail.to_csv('3.2.retail_transactions_preprocessed.csv')

df_retail.shape

# Narrow our focus to the transactions of 2019 only.

df_retail_19 = df_retail[df_retail['Year'].isin([2019])].copy()
df_retail_19['DaysSince'] = (dt.datetime(year=2019, month=12, day=31) - df_retail_19['InvoiceDate']).apply(lambda x: x.days)
df_retail_19.head()

# In this step we will create features for each individual customers.

operations = {
    'Revenue': 'sum',
    'DaysSince': ['max','min','nunique']
}
df_retail_19 = df_retail_19.groupby('CustomerID').agg(operations)
df_retail_19.columns = ['_'.join(col).strip() for col in df_retail_19.columns.values]
df_retail_19['AvgOrderCost'] = df_retail_19['Revenue_sum'] / df_retail_19['DaysSince_nunique']
df_retail_19


# Finally, we'll take the total revenue of a customer in 2020 as the label or dependent variable

revenue_2020 = df_retail[df_retail['Year'] == 2020].groupby('CustomerID')['Revenue'].sum()
revenue_2020.head()

wrangled_df = pd.concat([df_retail_19, revenue_2020], axis=1, join="inner")
wrangled_df = wrangled_df.rename(columns={
    'Revenue_sum': 'revenue_2019',
    'DaysSince_max': 'customer_relation_days',
    'DaysSince_min': 'last_purchase_days',
    'DaysSince_nunique': 'number_of_purchases',
    'AvgOrderCost': 'avg_order_cost',
    'Revenue': 'revenue_2020'})
wrangled_df.index.name = 'customer_id'

wrangled_df.to_csv('wrangled_transactions.csv')
wrangled_df.head()
