==> 2.5-Model-Building-and-Evaluation.ipynb <==
# import important libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px

from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, r2_score, root_mean_squared_error

# load the transaction dataset
df_retail = pd.read_csv('data/wrangled_transactions.csv')
df_retail.head()

### Model Building

X = df_retail[['revenue_2019', 'customer_relation_days', 'last_purchase_days', 'number_of_purchases', 'avg_order_cost']]
y = df_retail['revenue_2020']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=11)

# build the linear regression model and train it
model = LinearRegression()
model.fit(X_train, y_train)

# coefficients
model.coef_

### Evaluation 1

# predict single customer revenue 2020
test_customer = pd.DataFrame({
    'revenue_2019': [5000],
    'customer_relation_days': [200],
    'last_purchase_days': [10],
    'number_of_purchases': [20],
    'avg_order_cost': [1]
})

model.predict(test_customer)

### Evaluation 2

# predict
y_pred = model.predict(X_test)

# prepare a dataframe with actual value and prediction
predictions = pd.DataFrame({'y_test': y_test.array, 'y_pred': y_pred})
predictions = predictions[predictions['y_test'] < 10000]

# plot predictions against the actual value
fig = px.scatter(predictions, x=predictions.index, y=['y_test', 'y_pred'],
                 color_discrete_sequence=['blue', 'red'],
                 labels={'x': 'Index', 'value': 'Revenue'},
                 title='Actual vs. Predicted Revenue')
fig.show()

np.corrcoef(model.predict(X_test), y_test)

## Evaluation 3

print("Root mean squared error: %.2f" % root_mean_squared_error(y_test, model.predict(X_test)))

# Model with positively correlated features

X = df_retail[['revenue_2019', 'number_of_purchases']]
y = df_retail['revenue_2020']

# test train split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=11)

# build the linear regression model and train it
model = LinearRegression()
model.fit(X_train, y_train)

# predict
y_pred = model.predict(X_test)

# prepare a dataframe with actual value and prediction
predictions = pd.DataFrame({'y_test': y_test.array, 'y_pred': y_pred})
predictions = predictions[predictions['y_test'] < 10000]

fig = px.scatter(predictions, x=predictions.index, y=['y_test', 'y_pred'],
                 color_discrete_sequence=['blue', 'red'],
                 labels={'x': 'Index', 'value': 'Revenue'},
                 title='Actual vs. Predicted Revenue')
fig.show()

print("Root mean squared error: %.2f" % root_mean_squared_error(y_test, y_pred))
