==> 4.5-Churn-Prediction-LR.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import confusion_matrix

# Load dataset

df_telco = pd.read_csv('telco_customer_churn.csv', header=0)

# Identify categorical and numeric columns

df_telco.drop(columns=['CustomerID'], inplace=True)

# Store categorical column names
categorical = df_telco.nunique()[df_telco.nunique() < 5].keys().tolist()
categorical.remove('Churn')

# Store numerical column names
numerical = [x for x in df_telco.columns if x not in categorical]
numerical.remove('Churn')

### One hot encoding

df_telco_encoded = pd.get_dummies(data=df_telco, columns=categorical, drop_first=True)

### Remove redundant features

corr_cols = [
    'OnlineSecurity_No internet service',
    'OnlineBackup_No internet service',
    'DeviceProtection_No internet service',
    'TechSupport_No internet service',
    'StreamingTV_No internet service',
    'StreamingMovies_No internet service',
    'MultipleLines_No phone service'
]
df_telco_encoded.drop(columns=corr_cols, inplace=True)

# Logistic Regression

y = df_telco_encoded['Churn']
X = df_telco_encoded.drop(columns='Churn')

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=123)

# Initialize logistic regression classifier
logreg = LogisticRegression(solver='lbfgs', penalty='l2', max_iter=1000)

# Fit the classifier with training data
logreg.fit(X_train, y_train)

# Predict churn labels on testing data
y_pred = logreg.predict(X_test)

y_pred[5:15]

# Calculate accuracy score on testing data
accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred)
recall = recall_score(y_test, y_pred)

# Print test accuracy
print('Accuracy:', round(accuracy, 4))
print('Precision:', round(precision, 4))
print('Recall:', round(recall, 4))

### Confusion matrix

confusion_matrix(y_test, y_pred)
